Set up HTML, CSS, JS. Added basic HTML Structure
I initialized the HTML, CSS, and JS files, and linked them. For my CSS
here I am using a modern normalize with a few other touches courtesy of
Matt Brictson(attribution in README). The HTML structure here utilizes
a couple of headers, my footer, and a button. I've also added, inbetween
the second header and the button, a div which I can use DOM manipulation
to add the randomized content to. Next step is to add some styling.