Added Footer and updated Styling for Accessibility
My preferred footer is relatively simple but uses some text and an icon
link to refer to the github page for the project. As I've been learning
in The Odin Project however, there are often accessibility concerns that
developers often don't address. Anticipating this I looked into some
ways to make the icon link work better for more users. I therefore
incorporated a series of suggestions from Kitty Giraudel(attribution
in README). Mainly this involves adding attributes to the icon image
to change how it will be interacted with by users who primarily navigate
with a keyboard or screen reader. Then a span is added to ensure that
there is some text description that a screen reader can use. This is
then hidden visually with CSS.