Corrected Normalize Implementation, Added Styling
I forgot to actually install modern normalize and import it properly.
This has been fixed. I also added some basic styling to the project.
I may add in more color or other features later but for now I want to
get the scripting done so that the page is actually usable.