Finished Basic Script for Generating Random Chords
The chord generator script has been added. It utilizes two arrays, one
with the Keys (I did not include enharmonic keys for this part), and
another with the chord qualities. When the button is clicked, the
function to pick a random chord is called. The function takes a random
index from both of the arrays and combines them, then updates the text
content of the pargraph above the button. Because the function uses the
length of the arrays to "pick" a random chord, I can add more chord
qualities or keys as necessary. I may add functionality in the future to
change these settings, but for now I am going to stick with just the
most commonly used keys, and only the chord qualities that I can
currently play on command.